Inline OverlayPositionCalculator into OverlayManager
OverlayPositionCalculator was a singleton with three pure pixel-conversion
functions (calculateSizePx / calculateXPx / calculateYPx) called only from
OverlayManager.createLayoutParams(). Moving them as internal top-level
functions in OverlayManager.kt drops one file and one type-name to
navigate, while preserving testability — they remain pure functions
callable from JVM unit tests.

OverlayManagerTest is unchanged in structure; it just calls
calculateOverlaySizePx etc. directly instead of through the
OverlayPositionCalculator namespace.
